Module frame_benchmarking_pallet_pov::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

BoundedValue

A value with a MEL bound of 32 byte.

DoubleMap1M

Warning: Doc-Only

LargeValue

4MiB value.

LargeValue2

Warning: Doc-Only

Map1M

A map with a maximum of 1M entries.

Map16M

A map with a maximum of 16M entries.

UnboundedMap

Warning: Doc-Only

UnboundedMap2

Warning: Doc-Only

UnboundedMapTwox

Warning: Doc-Only

UnboundedValue

A value without a MEL bound.

Value

Warning: Doc-Only

Value2

Warning: Doc-Only